Use WindowCommands to reduce some hazzle
FileManager shouldn't abuse the ApplicationCommand class but use the
WindowCommand which is designed to be used for window related commands.

By doing so some of the hazzle in the different commands used to
retrieve the active window and/or file can be omitted.

It helps to make the commands a bit more pythonic.

Note:

1. This commit is a suggestion and may some review before merging.
2. It may break compatibility with Sublime Text 2.